refactor: move Slack types to their own file
Moves the Slack notifier's internal types (`Notifier`, `request`,
`attachment`, `slackResponse`) from slack.go into a new types.go.
No behavioral changes.

While moving them, two doc comments were tightened:
- `request`: "slack notification" -> "Slack notification"
- `attachment`: "richly-formatted message block" -> "richly formatted
  message block"
